nextjs-streamlit-oci-realtime-speech/pyproject.toml
[project]
name = "oracle-ai-speech"
version = "0.1.0"
description = "Real-time speech transcription using Oracle AI Speech services"
requires-python = ">=3.12"
dependencies = [
    "oci>=2.168.0",
    "streamlit>=1.55.0",
    "python-dotenv>=1.2.2",
    "sounddevice>=0.5.5",
    "numpy>=2.4.2",
    "pyaudio>=0.2.14",
    "oci-ai-speech-realtime>=2.2.0",
]

[tool.uv]
dev-dependencies = []
